Update Readme.md
Hey, I really liked the project and was reading through the Readme.md file when I came across some grammatical and spelling errors that you might have missed while editing the documentation. Also, I have improved some phrases or words that might be more suited to the documentation. It would be really a great opportunity for me if I could contribute to this project. Thank you.